fix(ui): align hover-timer field types with Bevy 0.18 Time::elapsed_secs (f32)
Time::elapsed_seconds() (f64) was renamed and retyped to elapsed_secs()
(f32) in Bevy 0.18. The hover-timer state in ShipbuildingUiState
(slot_hover_started_at, module_hover_started_at) and the
HOVER_LATENCY_SECS constant were f64, mismatching the new return type.

Switch the field types and constant to f32. The 250ms precision is
well within f32 range (a full game day of f32 ticks is more than
sufficient).

Refs: GRA-13

+++ src/ui/shipbuilding_state.rs
@@ -55,8 +55,8 @@ pub(crate) struct ShipbuildingUiState {
     pub hovered_slot: Option<String>,
     pub hovered_module_id: Option<String>,
     pub library_filter_query: String,
-    pub slot_hover_started_at: Option<f64>,
-    pub module_hover_started_at: Option<f64>,
+    pub slot_hover_started_at: Option<f32>,
+    pub module_hover_started_at: Option<f32>,
 }
 
 impl Default for ShipbuildingUiState {
